+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
  * re-degign classes.
+ * Support Email links.
 
 ## v0.0.0 (2023-08-24)